+++ GauriBusinessLogic/Costs.cs
@@ -11,9 +11,7 @@ namespace GauriBusinessLogic
     {
         public int Id { get; set; }
         public DateTime Date { get; set; }
-        public float Amount { get; set; }
+        public decimal Amount { get; set; }
         public string Description { get; set; }
-        //public DbSet<Costs> Costses { get; set; }
-
     }
 }
